restructuring app a bit
Caching works

renaming stuff
