+++ nom_track_app/config.py
@@ -1,0 +1,4 @@
+today_response_dict = {
+                        "date": "",
+                        "food_sources": [],
+                       }
