+++ requirements.txt
@@ -1,2 +1,3 @@
 flask==0.12.2
 requests==2.18.4
+Flask-Cache==0.13.1
